requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiohttp-jinja2==1.5.1
aiopg==1.4.0
aiopg-sqlite==0.0.1a0
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.5
cryptography==40.0.2
DateTime==5.1
Django==4.1.8
django-qrcode==0.3
django-smart-selects==1.6.0
fernet==1.0.1
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.17
pyotp==2.8.0
pypng==0.20220715.0
pytest==7.3.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.21.0
pytest-cov==4.0.0
pytz==2023.3
qrcode==7.4.2
SQLAlchemy==1.4.48
sqlparse==0.4.4
typing_extensions==4.5.0
Werkzeug==2.3.3
yarl==1.9.2
zope.interface==6.0
